fix(api): tri, pagination et projection PostgreSQL alignes sur MongoDB
Le tri suit desormais l'ordre BSON au lieu de refuser ou de trier a la
maniere SQL :
- texte en ordre binaire (COLLATE "C") quelle que soit la collation de la base ;
- NaN sous tous les nombres, absents en tete ascendant / en queue descendant ;
- tableaux par leur plus petit element (asc) ou leur plus grand (desc), le
  tableau vide avant l'absent, chemins dans les elements (lines.qty) inclus ;
- POJO embarque entier et map compares comme des sous-documents BSON (type,
  nom, valeur, champ par champ) via PgBsonOrder ;
- '_id' designe l'uuid ; un champ inconnu n'ajoute aucune cle (D6) ;
- l'uuid (COLLATE "C") termine l'ordre des qu'une page est demandee (D5).

L'offset est calcule en long (PgQuery.offset devient Long) : une page au-dela
des donnees est vide au lieu d'un debordement. Une projection pointee ne garde
que son segment de tete, comme MongoDao.applyProjection.

Nouveaux tests : PgSortPageTest (forme SQL et ordre reel) et
ParitySortShapesTest (10 scenarios de parite sur tableaux, sous-documents,
collections de POJO et maps).

+++ api/dao/dao-postgresql/src/main/java/com/garganttua/dao/postgresql/PgQuery.java
@@ -16,10 +16,10 @@ import java.util.Objects;
  * @param params     the bind parameters of {@code where}, in order, already JDBC-ready
  * @param orderBy    the {@code ORDER BY …} clause including the keyword, or an empty string
  * @param limit      the page size, or null for no limit
- * @param offset     the rows to skip, or null
+ * @param offset     the rows to skip, or null — a {@code long}: {@code index * size} may exceed an int
  * @param projection the DTO field paths to load (dotted), or null to load everything
  */
-public record PgQuery(String where, List<Object> params, String orderBy, Integer limit, Integer offset,
+public record PgQuery(String where, List<Object> params, String orderBy, Integer limit, Long offset,
         List<String> projection) {
 
     /** The alias of the main table in every generated statement. */
